+++ source/extensions/omni.isaac.ros_bridge/python/nodes/OgnROS1RtxLidarHelper.py
@@ -43,43 +43,43 @@ class OgnROS1RtxLidarHelper:
     @staticmethod
     def compute(db) -> bool:
         if db.inputs.enabled is False:
-            if db.internal_state.initialized is False:
+            if db.per_instance_state.initialized is False:
                 return True
             else:
-                db.internal_state.custom_reset()
+                db.per_instance_state.custom_reset()
                 return True
-        if db.internal_state.initialized is False:
-            db.internal_state.initialized = True
+        if db.per_instance_state.initialized is False:
+            db.per_instance_state.initialized = True
             stage = omni.usd.get_context().get_stage()
             keys = og.Controller.Keys
             with Usd.EditContext(stage, stage.GetSessionLayer()):
                 render_product_path = db.inputs.renderProductPath
                 if not render_product_path:
                     carb.log_warn("Render product not valid")
-                    db.internal_state.initialized = False
+                    db.per_instance_state.initialized = False
                     return False
                 if stage.GetPrimAtPath(render_product_path) is None:
                     # Invalid Render Product Path
                     carb.log_warn("Render product not created yet, retrying on next call")
-                    db.internal_state.initialized = False
+                    db.per_instance_state.initialized = False
                     return False
                 else:
                     prim = stage.GetPrimAtPath(get_camera_prim_path(render_product_path))
                     if prim.IsA(UsdGeom.Camera):
                         if prim.HasAPI(IsaacSensorSchema.IsaacRtxLidarSensorAPI):
-                            db.internal_state.render_product_path = render_product_path
-                            db.internal_state.sensor = "lidar"
+                            db.per_instance_state.render_product_path = render_product_path
+                            db.per_instance_state.sensor = "lidar"
                         else:
-                            db.internal_state.sensor = None
+                            db.per_instance_state.sensor = None
 
-                if db.internal_state.sensor is None:
+                if db.per_instance_state.sensor is None:
                     carb.log_warn("Active camera for Render product is not an RTX Lidar")
-                    db.internal_state.initialized = False
+                    db.per_instance_state.initialized = False
                     return False
 
-                db.internal_state.render_product_path = render_product_path
+                db.per_instance_state.render_product_path = render_product_path
                 sensor_type = db.inputs.type
-                db.internal_state.resetSimulationTimeOnStop = db.inputs.resetSimulationTimeOnStop
+                db.per_instance_state.resetSimulationTimeOnStop = db.inputs.resetSimulationTimeOnStop
                 writer = None
                 try:
                     if sensor_type == "laser_scan":
@@ -95,7 +95,7 @@ class OgnROS1RtxLidarHelper:
 
                     else:
                         carb.log_error("type is not supported")
-                        db.internal_state.initialized = False
+                        db.per_instance_state.initialized = False
                         return False
                     if writer is not None:
                         writer.initialize(
@@ -104,11 +104,11 @@ class OgnROS1RtxLidarHelper:
                             queueSize=db.inputs.queueSize,
                             topicName=db.inputs.topicName,
                         )
-                        db.internal_state.append_writer(writer)
+                        db.per_instance_state.append_writer(writer)
                         if db.inputs.showDebugView:
                             writer = rep.writers.get("RtxLidar" + "DebugDrawPointCloud" + "Buffer")
-                            db.internal_state.append_writer(writer)
-                    db.internal_state.attach_writers(render_product_path)
+                            db.per_instance_state.append_writer(writer)
+                    db.per_instance_state.attach_writers(render_product_path)
                 except Exception as e:
                     print(traceback.format_exc())
                     pass
@@ -116,9 +116,9 @@ class OgnROS1RtxLidarHelper:
             return True
 
     @staticmethod
-    def release(node):
+    def release_instance(node, graph_instance_id):
         try:
-            state = OgnROS1RtxLidarHelperInternalState.per_node_internal_state(node)
+            state = OgnROS1RtxLidarHelperInternalState.per_instance_internal_state(node)
         except Exception:
             state = None
             pass